+++ resources/[local]/v-core/fxmanifest.lua
@@ -5,7 +5,7 @@ lua54 'yes'
 name 'v-core'
 author 'vyrriox'
 description 'v-core — custom roleplay framework core'
-version '0.1.0'
+version '0.2.0'
 
 shared_scripts {
     'config/config.lua',
@@ -13,12 +13,16 @@ shared_scripts {
 }
 
 client_scripts {
+    'client/callbacks.lua',
     'client/functions.lua',
     'client/main.lua',
 }
 
 server_scripts {
+    '@oxmysql/lib/MySQL.lua',
     'server/functions.lua',
+    'server/database.lua',
+    'server/callbacks.lua',
     'server/player.lua',
     'server/main.lua',
 }